+++ tmux-rs/src/cmd_/cmd_list_clients.rs
@@ -39,7 +39,7 @@ unsafe extern "C" fn cmd_list_clients_exec(self_: *mut cmd, item: *mut cmdq_item
                 continue;
             }
 
-            let mut ft = format_create(cmdq_get_client(item), item, FORMAT_NONE as i32, format_flags::empty());
+            let mut ft = format_create(cmdq_get_client(item), item, FORMAT_NONE, format_flags::empty());
             format_add(ft, c"line".as_ptr(), c"%u".as_ptr(), idx);
             format_defaults(ft, c, None, None, None);
