+++ tmux-rs/src/cmd_/cmd_source_file.rs
@@ -134,7 +134,7 @@ unsafe extern "C" fn cmd_source_file_exec(self_: *mut cmd, item: *mut cmdq_item)
             (*cdata).flags |= cmd_parse_input_flags::CMD_PARSE_VERBOSE;
         }
 
-        utf8_stravis(&raw mut cwd, server_client_get_cwd(c, null_mut()), VIS_GLOB as i32);
+        utf8_stravis(&raw mut cwd, server_client_get_cwd(c, null_mut()), VIS_GLOB);
 
         for i in 0..args_count(args) {
             let mut path = args_string(args, i);
